second impl pass
